Refactor player shell: trim puzzle-export barrel, split shift/leaderboard
- Import omitEmptyNextLetterSlots from puzzle-export-sim/next-letters (avoid barrel on player load).

- Extract game-player-dom-refs, game-player-shell, preloadGameSoundLayers from game.js.

- Add leaderboard-client for live leaderboard fetch/POST + JSON normalization.

- Split shift preview DOM vs gesture controller; shift-dom re-exports for stable imports.

- Drop unused imports in game.js.
