+++ js/game-player-dom-refs.js
@@ -1,0 +1,40 @@
+/** @param {Record<string, unknown>} refs Destination bag (typically `ctx.refs`). */
+export function assignGamePlayerDomRefs(refs) {
+  Object.assign(refs, {
+    grid: document.querySelector("#grid"),
+    gridPan: document.getElementById("grid-pan"),
+    gridStage: document.getElementById("grid-stage"),
+    shiftPreviewStrip: document.getElementById("shift-preview-strip"),
+    startButton: document.querySelector("#start"),
+    currentWordElement: document.querySelector("#current-word"),
+    queueNextHeaderElement: document.querySelector("#queue-next-header"),
+    nextLettersElement: document.querySelector("#queue-next-values"),
+    queueSackCountElement: document.querySelector("#queue-sack-count"),
+    scoreElement: document.querySelector("#score"),
+    scoreSwipeSumElement: document.querySelector("#score-swipe-sum"),
+    scoreLengthElement: document.querySelector("#score-length"),
+    scoreWordTotalElement: document.querySelector("#score-word-total"),
+    scoreGameTotalElement: document.querySelector("#score-game-total"),
+    gameInfoContainer: document.querySelector("#game-info-container"),
+    bottomDock: document.querySelector("#bottom-dock"),
+    rules: document.querySelector("#rules"),
+    rulesButton: document.querySelector("#rules-button"),
+    rulesPerfectHuntTotalElement: document.querySelector("#rules-perfect-hunt-total"),
+    rulesPerfectHuntCountElement: document.querySelector("#rules-perfect-hunt-count"),
+    rulesNextLettersCountElement: document.querySelector("#rules-next-letters-count"),
+    muteButton: document.getElementById("mute-button"),
+    doneButton: document.querySelector("#done-button"),
+    boardShiftZone: document.getElementById("board-shift-zone"),
+    boardShiftHints: document.getElementById("board-shift-hints"),
+    buttonContainer: document.getElementById("button-container"),
+    retryButton: document.querySelector("#retry-button"),
+    gridLineContainer: document.querySelector("#line-container"),
+    gridLineWrapper: document.getElementById("grid-line-wrapper"),
+    gridViewport: document.getElementById("grid-viewport"),
+    leaderboardElements: document.getElementById("leaderboard-elements"),
+    leaderboardTable: document.getElementById("leaderboard-table"),
+    playerName: document.getElementById("player-name"),
+    leaderboardButton: document.getElementById("leaderboard-button"),
+    leaderboardDemoAdd: document.getElementById("leaderboard-demo-add"),
+  });
+}
